add relative X/Y for nicer front-end rendering
this ensures that the browser (smkiwi) can render it nicely
relative to the location pin without any calculations client-side

I've also refactored coords from using doubles to floats as float precision
is fine

+++ Satellite.Models/CurrentCoords.cs
@@ -2,7 +2,7 @@
 {
     public class CurrentCoords
     {
-        public double Longitude;
-        public double Latitude;
+        public float Longitude;
+        public float Latitude;
     }
 }
